refactor: relocate helper function
